typeshare用のアノテーション追加、id関連の型をi64->i32へ変更、コード修正

+++ src-tauri/src/commands/tag.rs
@@ -1,10 +1,10 @@
 use sqlx::{Pool, Sqlite};
 
-pub async fn create_tag_in_db(sqlite_pool: Pool<Sqlite>, name: String) -> Result<i64, ()> {
+pub async fn create_tag_in_db(sqlite_pool: Pool<Sqlite>, name: String) -> Result<i32, ()> {
     let tag = sqlx::query("INSERT INTO Tags (name) VALUES (?)")
         .bind(name)
         .execute(&sqlite_pool)
         .await
         .unwrap();
-    Ok(tag.last_insert_rowid())
+    Ok(tag.last_insert_rowid() as i32)
 }
